Python/requirements.txt
flask
transformers
torch
faiss-cpu
numpy